# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Linha",
    "version": "0.1.0",
    "description": "Versioned JSON base protocol. SDKs wrap these operations. Result templates and policies are snapshotted; job IDs survive process restarts. Authentication follows deployment configuration: security.enabled=false disables client and worker authentication; security.oidc.enabled=false disables only client authentication. Anonymous clients share owner anonymous. Secured deployments require the bearer schemes below."
  },
  "paths": {
...
    "/v1/admin/session": {
      "get": {
        "operationId": "getAdminV1AdminSession",
        "summary": "Read administrator identity and CSRF token",
        "security": [
          {
            "adminSession": []
          },
          {
            "clientBearer": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSession"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/logout": {
      "post": {
        "operationId": "postAdminV1AdminLogout",
        "summary": "End the browser session (viewer or operator)",
        "security": [
          {
            "adminSession": []
          },
          {
            "clientBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Linha-CSRF",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Required with same-origin browser-session mutations; bearer requests use verified token authorization."
          },
          {
            "name": "Origin",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/contexts": {
      "get": {
        "operationId": "getAdminV1AdminContexts",
        "summary": "Read immutable versions without creating demand",
        "security": [
          {
            "adminSession": []
          },
          {
            "clientBearer": []
          }
        ],
        "parameters": [
          {
            "name": "owner",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "engine",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contextId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "metricContext",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          },
          {
            "name": "includeStopped",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "until",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalContextPage"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/contexts/{context}": {
      "get": {
        "operationId": "getAdminV1AdminContextsContext",
        "summary": "Read one context observation",
        "security": [
          {
            "adminSession": []
          },
          {
            "clientBearer": []
          }
        ],
        "parameters": [
          {
            "name": "context",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalContext"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/admin/servers": {
      "get": {
        "operationId": "getAdminV1AdminServers",
        "summary": "Read recent server heartbeats",
        "security": [
          {
            "adminSession": []
          },
          {
            "clientBearer": []
          }
        ],
        "parameters": [
          {
            "name": "owner",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "engine",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contextId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "metricContext",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          },
          {
            "name": "includeStopped",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "until",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalServerPage"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/jobs": {
      "get": {
        "operationId": "getAdminV1AdminJobs",
        "summary": "Read requests across owners with bounded filters",
        "security": [
          {
            "adminSession": []
          },
          {
            "clientBearer": []
          }
        ],
        "parameters": [
          {
            "name": "owner",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "engine",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contextId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "metricContext",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          },
          {
            "name": "includeStopped",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "until",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminJobPage"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/jobs/{job}": {
      "get": {
        "operationId": "getAdminV1AdminJobsJob",
        "summary": "Read immutable request input, status and provenance",
        "security": [
          {
            "adminSession": []
          },
          {
            "clientBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminJob"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/jobs/{job}/audit": {
      "get": {
        "operationId": "getAdminV1AdminJobsAuditJob",
        "summary": "Read accepted administrative audit",
        "security": [
          {
            "adminSession": []
          },
          {
            "clientBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "owner",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "engine",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contextId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "metricContext",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          },
          {
            "name": "includeStopped",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "until",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAuditPage"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/contexts/{context}/jobs": {
      "post": {
        "operationId": "postAdminV1AdminContextsJobsContext",
        "summary": "Accept an administrative create or replay transaction",
        "security": [
          {
            "adminSession": []
          },
          {
            "clientBearer": []
          }
        ],
        "parameters": [
          {
            "name": "context",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Linha-CSRF",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Required with same-origin browser-session mutations; bearer requests use verified token authorization."
          },
          {
            "name": "Origin",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Committed acceptance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminSubmitRequest"
              }
            }
          }
        }
      }
    },
    "/v1/admin/jobs/{job}/cancel": {
      "post": {
        "operationId": "postAdminV1AdminJobsCancelJob",
        "summary": "Cancel one job with existing attempt fencing",
        "security": [
          {
            "adminSession": []
          },
          {
            "clientBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Linha-CSRF",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Required with same-origin browser-session mutations; bearer requests use verified token authorization."
          },
          {
            "name": "Origin",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/admin/jobs/{job}/preview": {
      "get": {
        "operationId": "getAdminV1AdminJobsPreviewJob",
        "summary": "Read an escaped bounded published JSON preview",
        "security": [
          {
            "adminSession": []
          },
          {
            "clientBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxBytes",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1048576,
              "default": 65536
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPreview"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/admin/jobs/{job}/files/{output}": {
      "get": {
        "operationId": "getAdminJobFilesoutput",
        "summary": "downloadResultFile",
        "security": [
          {
            "adminSession": []
          },
          {
            "clientBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized streamed file bytes",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "job",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "output",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/ui/oidc/login": {
      "get": {
        "summary": "OIDC code/PKCE browser login; callback requires single-use state cookie and nonce",
        "security": [],
        "responses": {
          "303": {
            "description": "Browser redirect"
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/ui/oidc/callback": {
      "get": {
        "summary": "OIDC code/PKCE browser login; callback requires single-use state cookie and nonce",
        "security": [],
        "responses": {
          "303": {
            "description": "Browser redirect"
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/admin/overview": {
      "get": {
        "operationId": "getAdminOverview",
        "summary": "Read complete operational totals and bounded recent detail",
        "security": [
          {
            "adminSession": []
          },
          {
            "clientBearer": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOverview"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AdminSession": {
        "type": "object",
        "properties": {
          "actor": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "viewer",
              "operator"
            ]
          },
          "csrf": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "actor",
          "role"
        ]
      },
...
    },
    "securitySchemes": {
      "clientBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Validated OIDC token; explicitly configured development token only in development mode"
      },
      "workerBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Credential scoped to worker context and incarnation"
      },
      "adminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "linha_session",
        "description": "Opaque HA browser session; viewer/operator mapping checked for each admin request. Cookie mutations additionally require Origin and X-Linha-CSRF."
      }
    }
  }
}